+++ app/assets/javascripts/thredded/dependencies/autosize.js
@@ -1,1 +1,1 @@
-//= require autosize.min
+//= require autosize