+++ spec/dummy/app/javascript/packs/application.js
@@ -1,0 +1,4 @@
+require('@rails/ujs').start();
+require('turbolinks').start();
+require('monkey-patch-turbolinks.js');
+require('thredded_imports.js');