+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+* Experimental Webpack support.
+  [#838](https://github.com/thredded/thredded/pull/838)
 * Sprockets 4.0.0 support.
 * Messageboard group page. Can be disabled by setting `Thredded.show_messageboard_group_page = false`.
   [#829](https://github.com/thredded/thredded/pull/829)